More updates

Former-commit-id: 17a9e5d045b6dec31e14441f596e7e4483596ee9 [formerly 3529a2667611ddeb81bb7245c82fa9774ba6e5be]
Former-commit-id: 9bd02fce5b09251e8abc25bef4d157e926bdc1c8

+++ lagom/transform/polyak_average.py
@@ -23,7 +23,7 @@ class PolyakAverage(object):
         Args:
             x (object): additional data to update the estimation of running average. 
         """
-        x = np.asarray(x, dtype=np.float32)
+        x = np.asarray(x)
         if self.x is None:
             self.x = x
         else:
